Feature | Working schedule task to send email |
This schedule task allows to send email with new offers to all users in database
Every 4 hours (6x a day) the email is sent to all users
