--- README.md
@@ -1,2 +1,0 @@
-# OLX-analitics
-A program that allows automatic selection of categories, filters, locations on the OLX portal, data collection and analysis
